requirements.txt
pandas
numpy
matplotlib
seaborn
openpyxl
requests
beautifulsoup4
click
pyinstaller
scipy
python-dotenv
pytest
pytest-cov
black
flake8
mypy
google-cloud-bigquery
google-auth
google-auth-oauthlib
dbfread